Initiated Data Analysis
Visually looked if there is a trend between experience levle and played hours through a scatterplot. Ie: we want to identify which groups would give the most amount of data. And started to find ways to test if age can predict if the player will give the most data for researches in terms of play time.